node_hub/lerobot_record/requirements.txt
dora-rs==0.3.5
numpy==2.0.0
pyarrow==16.1.0

opencv-python==4.10.0.84
pygame==2.6.0